requirements.txt
fastapi[standard]>=0.110.0,<0.114.0
pydantic>=2.7.0,<3.0.0
robotframework>=7.0.1
motor~=3.3
uvicorn~=0.28
pydantic[email]
